Fix mesh indexing. Fix raytrace closest hit shader. Improve shaders.
- Fixed mesh indexing. Uses vectors now instead of map to store mesh data.
- Fixed closest hit shader because of change in mesh indexing.
- Removed unecessary buffer data sent to shaders.
- Added inverse view-projection matrix to camera buffer.
- Improved raster shaders.

+++ kirana/viewport/vulkan/command_buffers.cpp
@@ -266,7 +266,7 @@ void kirana::viewport::vulkan::CommandBuffers::end(uint32_t index) const
 void kirana::viewport::vulkan::CommandBuffers::buildAccelerationStructure(
     const vk::AccelerationStructureBuildGeometryInfoKHR &geoInfo,
     const vk::AccelerationStructureBuildRangeInfoKHR *rangeInfo,
-    vk::QueryPool &compactionPool, size_t firstCompaction,
+    vk::QueryPool &compactionPool, uint32_t firstCompaction,
     bool addMemoryBarrier, uint32_t index) const
 {
     m_current[index].buildAccelerationStructuresKHR(geoInfo, rangeInfo);
